requirements.txt
openai==1.3.9
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
PyPDF2==3.0.1
python-docx==0.8.11
langchain==0.1.0
langchain-openai==0.0.5
